elfowl/requirements.txt
## --- elfowl requirements --- ##
openai==1.12.0
tiktoken==0.6.0
requests==2.31.0
pandas==2.2.1
PyGithub==2.2.0
argparse==1.4.0
numpy==1.26.4
langchain==0.1.9
backoff==2.2.1
ply==3.11
py2cfg==0.7.3
django==5.0.3
python-decouple==3.8
